refactor(utils): move FinalizeDownload to its own package
