--- src/Main.js
@@ -1,7 +1,0 @@
-/**
- * Created by tusharmathur on 5/15/15.
- */
-'use strict'
-import * as U from './Utils'
-
-export const UTILS = U
